# locationAPI/server/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "The ETSI MEC ISG MEC012 Location API described using OpenAPI. The\
    \ API is based on the Open Mobile Alliance's specification RESTful Network API\
    \ for Zonal Presence"
  version: "1.0.0"
  title: "Location API"
  license:
    name: "BSD-3-Clause"
    url: "https://forge.etsi.org/legal-matters"
host: "127.0.0.1:8081"
basePath: "/exampleAPI/location/v1"
...
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /subscriptions/zonalTraffic:
    get:
      tags:
      - "subscriptions"
      description: "This operation is used for retrieving all active subscriptions\
        \ to zonal traffic change notifications."
      operationId: "zonal_traffic_sub_get"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Response to retrieve zonal traffic subscriptions"
          examples:
            application/json:
              notificationSubscriptionList:
                zonalTrafficSubscription:
                - clientCorrelator: "0123"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zonalTraffic/subscription123"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123456"
                  zoneId: "zone01"
                  interestRealm: "LA"
                  userEventCriteria: "Transferring"
                - clientCorrelator: "0124"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zonalTraffic/subscription124"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123457"
                  zoneId: "zone02"
                  interestRealm: "LA"
                  userEventCriteria: "Transferring"
                resourceURL: "http://example.com/exampleAPI/location/v1/zonalTraffic"
          schema:
            $ref: "#/definitions/inline_response_200"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
    post:
      tags:
      - "subscriptions"
      description: "This operation is used for creating a new subscription to zonal\
        \ traffic change notification."
      operationId: "zonal_traffic_sub_post"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "zonalTrafficSubscription"
        description: "Zonal Traffic Subscription"
        required: true
        schema:
          $ref: "#/definitions/ZonalTrafficSubscription"
      responses:
        "201":
          description: "Response to create new zonal traffic subscription"
          examples:
            application/json:
              zonalTrafficSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zonalTraffic/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                zoneId: "zone01"
                interestRealm: "LA"
                userEventCriteria: "Transferring"
          schema:
            properties:
              zonalTrafficSubscription:
                $ref: "#/definitions/ZonalTrafficSubscription"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
  /subscriptions/zonalTraffic/{subscriptionId}:
    get:
      tags:
      - "subscriptions"
      description: "This operation is used for updating an individual subscription\
        \ to zonal traffic change notification."
      operationId: "zonal_traffic_sub_get_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Response to retrieve individual zonal traffic subscription"
          examples:
            application/json:
              zonalTrafficSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zonalTraffic/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                zoneId: "zone01"
                interestRealm: "LA"
                userEventCriteria: "Transferring"
          schema:
            properties:
              zonalTrafficSubscription:
                $ref: "#/definitions/ZonalTrafficSubscription"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
    put:
      tags:
      - "subscriptions"
      description: "This operation is used for updating an individual subscription\
        \ to zonal traffic change notification."
      operationId: "zonal_traffic_sub_put_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      - in: "body"
        name: "zonalTrafficSubscription"
        description: "Zonal Traffic Subscription"
        required: true
        schema:
          $ref: "#/definitions/ZonalTrafficSubscription"
      responses:
        "200":
          description: "Response to update individual zonal traffic subscription"
          examples:
            application/json:
              zonalTrafficSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zonalTraffic/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                zoneId: "zone01"
                interestRealm: "LA"
                userEventCriteria: "Transferring"
          schema:
            properties:
              zonalTrafficSubscription:
                $ref: "#/definitions/ZonalTrafficSubscription"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
    delete:
      tags:
      - "subscriptions"
      description: "This operation is used for cancelling a subscription and stopping\
        \ corresponding notifications."
      operationId: "zonal_traffic_sub_del_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      responses:
        "204":
          description: "No content"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
  /subscriptions/userTracking:
    get:
      tags:
      - "subscriptions"
      description: "This operation is used for retrieving all active subscriptions\
        \ to user tracking change notifications."
      operationId: "user_tracking_sub_get"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Response to retrieve user tracking subscriptions"
          examples:
            application/json:
              notificationSubscriptionList:
                userTrackingSubscription:
                - clientCorrelator: "0123"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123456"
                  address: "acr:192.0.2.1"
                  userEventCriteria: "Transferring"
                - clientCorrelator: "0124"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription124"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123456"
                  address: "acr:192.0.2.2"
                  userEventCriteria: "Transferring"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking"
          schema:
            $ref: "#/definitions/inline_response_200_1"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
    post:
      tags:
      - "subscriptions"
      description: "This operation is used for creating a new subscription to user\
        \ tracking change notification"
      operationId: "user_tracking_sub_post"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "userTrackingSubscription"
        description: "User Tracking Subscription"
        required: true
        schema:
          $ref: "#/definitions/UserTrackingSubscription"
      responses:
        "201":
          description: "Response to create new user tracking subscription"
          examples:
            application/json:
              userTrackingSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                address: "acr:192.0.2.1"
                userEventCriteria: "Transferring"
          schema:
            properties:
              userTrackingSubscription:
                $ref: "#/definitions/UserTrackingSubscription"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
  /subscriptions/userTracking/{subscriptionId}:
    get:
      tags:
      - "subscriptions"
      description: "This operation is used for retrieving an individual subscription\
        \ to user tracking change notification."
      operationId: "user_tracking_sub_get_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Response to retrieve individual user tracking subscription"
          examples:
            application/json:
              userTrackingSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                address: "acr:192.0.2.1"
                userEventCriteria: "Transferring"
          schema:
            properties:
              userTrackingSubscription:
                $ref: "#/definitions/UserTrackingSubscription"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
    put:
      tags:
      - "subscriptions"
      description: "This operation is used for updating an individual subscription\
        \ to user tracking change notification."
      operationId: "user_tracking_sub_put_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      - in: "body"
        name: "userTrackingSubscription"
        description: "User Tracking Subscription"
        required: true
        schema:
          $ref: "#/definitions/UserTrackingSubscription"
      responses:
        "200":
          description: "Response to update individual user tracking subscription"
          examples:
            application/json:
              userTrackingSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                address: "acr:192.0.2.1"
                userEventCriteria: "Transferring"
          schema:
            properties:
              userTrackingSubscription:
                $ref: "#/definitions/UserTrackingSubscription"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
    delete:
      tags:
      - "subscriptions"
      description: "This operation is used for retrieving an individual subscription\
        \ to user tracking change notification."
      operationId: "user_tracking_sub_del_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      responses:
        "204":
          description: "No Content"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
  /subscriptions/zonalStatus:
    get:
      tags:
      - "subscriptions"
      description: "This operation is used for creating a new subscription to zone\
        \ status change notification."
      operationId: "zone_status_get"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Response to retrieve zone status subscriptions"
          examples:
            application/json:
              notificationSubscriptionList:
                zoneStatusSubscription:
                - clientCorrelator: "0123"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zoneStatus/subscription123"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123456"
                  zoneId: "zone01"
                  numberOfUsersZoneThreshold: "500"
                  operationStatus: "Serviceable"
                - clientCorrelator: "0124"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zoneStatus/subscription124"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123457"
                  zoneId: "zone02"
                  numberOfUsersAPThreshold: "50"
                  operationStatus: "Serviceable"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zoneStatus"
          schema:
            $ref: "#/definitions/inline_response_200_2"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
    post:
      tags:
      - "subscriptions"
      description: "This operation is used for creating a new subscription to zone\
        \ status change notification."
      operationId: "zone_status_post"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "zoneStatusSubscription"
        description: "Zone Status Subscription"
        required: true
        schema:
          $ref: "#/definitions/ZoneStatusSubscription"
      responses:
        "201":
          description: "Response to create new zone status subscription"
          examples:
            application/json:
              zoneStatusSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zoneStatus/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                zoneId: "zone01"
                numberOfUsersZoneThreshold: "500"
                operationStatus: "Serviceable"
          schema:
            properties:
              zonalTrafficSubscription:
                $ref: "#/definitions/ZoneStatusSubscription"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
  /subscriptions/zoneStatus/{subscriptionId}:
    get:
      tags:
      - "subscriptions"
      description: "This operation is used for retrieving an individual subscription\
        \ to zone status change notification."
      operationId: "zone_status_get_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Response to retrieve individual zone status subscription"
          examples:
            application/json:
              zoneStatusSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zoneStatus/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                zoneId: "zone01"
                numberOfUsersZoneThreshold: "500"
                operationStatus: "Serviceable"
          schema:
            properties:
              zoneStatusSubscription:
                $ref: "#/definitions/ZoneStatusSubscription"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
    put:
      tags:
      - "subscriptions"
      description: "This operation is used for updating an individual subscription\
        \ to zone status change notification."
      operationId: "zone_status_put_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      - in: "body"
        name: "zoneStatusSubscription"
        description: "Zone Status Subscription"
        required: true
        schema:
          $ref: "#/definitions/ZoneStatusSubscription"
      responses:
        "200":
          description: "Response to update individual zone status subscription"
          examples:
            application/json:
              zoneStatusSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zoneStatus/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                zoneId: "zone01"
                numberOfUsersZoneThreshold: "500"
                operationStatus: "Serviceable"
          schema:
            properties:
              zoneStatusSubscription:
                $ref: "#/definitions/ZoneStatusSubscription"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
    delete:
      tags:
      - "subscriptions"
      description: "This operation is used for cancelling a subscription and stopping\
        \ corresponding notifications."
      operationId: "zone_status_del_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      responses:
        "204":
          description: "No content"
      x-swagger-router-controller: "swagger_server.controllers.subscriptions_controller"
definitions:
...
  Address:
    type: "string"
    format: "uri"
    description: "Address of user (e.g. \"sip\" URI, \"tel\" URI, \"acr\" URI)."
    example: "acr:192.0.2.1"
...
  CallbackReference:
    type: "object"
    required:
    - "notifyURL"
    properties:
      notifyURL:
        $ref: "#/definitions/NotifyURL"
    description: "Notification callback definition."
    example:
      notifyURL: "http://clientApp.example.com/location_notifications/123456"
  ClientCorrelator:
    type: "string"
    description: "Uniquely identifies this create subscription request. If there is\
      \ a communication failure during the request, using the same clientCorrelator\
      \ when retrying the request allows the operator to avoid creating a duplicate\
      \ subscription."
    example: "0123"
...
  Duration:
    type: "string"
    description: "Period (in seconds) of time notifications are provided for. If set\
      \ to \"0\" (zero), a default duration time, which is specified by the service\
      \ policy, will be used. If the parameter is omitted, the notifications will\
      \ continue until the maximum duration time, which is specified by the service\
      \ policy, unless the notifications are stopped by deletion of subscription for\
      \ notifications. This element MAY be given by the client during resource creation\
      \ in order to signal the desired lifetime of the subscription. The server MUST\
      \ return in this element the period of time for which the subscription will\
      \ still be valid."
    example: "0"
  InterestRealm:
    type: "string"
    description: "Interest realm of access point (e.g. geographical area, a type of\
      \ industry etc.)."
    example: "LA"
...
  NotifyURL:
    type: "string"
    format: "url"
    description: "The URL of your own listener application."
    example: "http://clientApp.example.com/location_notifications/123456"
...
  NumberOfUsersAPThreshold:
    type: "integer"
    format: "uint32"
    description: "Threshold number of users in an access point which if crossed shall\
      \ cause a notification."
    example: "20"
...
  NumberOfUsersZoneThreshold:
    type: "integer"
    format: "uint32"
    description: "Threshold number of users in a zone which if crossed shall cause\
      \ a notification."
    example: "40"
  OperationStatus:
    type: "string"
    description: "The operation status of the access point"
    example: "Serviceable"
    enum:
    - "Serviceable"
    - "Unserviceable"
    - "Unknown"
...
  ResourceURL:
    type: "string"
    format: "uri"
    description: "Self referring URL."
    example: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
...
  UserEventCriteria:
    type: "array"
    description: "List of user event values to generate notifications for (these apply\
      \ to address specified). If this element is missing, a notification is requested\
      \ to be generated for any change in user event."
    items:
      $ref: "#/definitions/UserEventType"
  UserEventType:
    type: "string"
    description: "User event"
    example: "Entering"
    enum:
    - "Entering"
    - "Leaving"
    - "Transferring"
...
  UserTrackingSubscription:
    type: "object"
    required:
    - "address"
    - "callbackReference"
    properties:
      clientCorrelator:
        $ref: "#/definitions/ClientCorrelator"
      callbackReference:
        $ref: "#/definitions/CallbackReference"
      address:
        $ref: "#/definitions/Address"
      userEventCriteria:
        $ref: "#/definitions/UserEventCriteria"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
    description: "A type containing user tracking subscription."
    example:
      address: "acr:192.0.2.1"
      resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
      callbackReference:
        notifyURL: "http://clientApp.example.com/location_notifications/123456"
      clientCorrelator: "0123"
      userEventCriteria: ""
...
  ZonalTrafficSubscription:
    type: "object"
    required:
    - "callbackReference"
    - "zoneId"
    properties:
      clientCorrelator:
        $ref: "#/definitions/ClientCorrelator"
      callbackReference:
        $ref: "#/definitions/CallbackReference"
      zoneId:
        $ref: "#/definitions/ZoneId"
      interestRealm:
        type: "array"
        description: "Interest realms of access points within a zone (e.g. geographical\
          \ area, a type of industry etc.)."
        items:
          $ref: "#/definitions/InterestRealm"
      userEventCriteria:
        type: "array"
        description: "List of user event values to generate notifications for (these\
          \ apply to zone identifier or all interest realms within zone identifier\
          \ specified). If this element is missing, a notification is requested to\
          \ be generated for any change in user event."
        items:
          $ref: "#/definitions/UserEventType"
      duration:
        $ref: "#/definitions/Duration"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
    description: "A type containing zonal traffic subscription"
    example:
      duration: "0"
      resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
      callbackReference:
        notifyURL: "http://clientApp.example.com/location_notifications/123456"
      clientCorrelator: "0123"
      zoneId: "zone01"
      interestRealm:
      - "LA"
      - "LA"
      userEventCriteria:
      - "Entering"
      - "Entering"
  ZoneId:
    type: "string"
    description: "Identifier of zone"
    example: "zone01"
...
  ZoneStatusSubscription:
    type: "object"
    required:
    - "callbackReference"
    - "zoneId"
    properties:
      clientCorrelator:
        $ref: "#/definitions/ClientCorrelator"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
      callbackReference:
        $ref: "#/definitions/CallbackReference"
      zoneId:
        $ref: "#/definitions/ZoneId"
      numberOfUsersZoneThreshold:
        $ref: "#/definitions/NumberOfUsersZoneThreshold"
      numberOfUsersAPThreshold:
        $ref: "#/definitions/NumberOfUsersAPThreshold"
      operationStatus:
        type: "array"
        description: "List of operation status values to generate notifications for\
          \ (these apply to all access points within a zone)."
        items:
          $ref: "#/definitions/OperationStatus"
    description: "A type containing zone status subscription."
    example:
      operationStatus:
      - "Serviceable"
      - "Serviceable"
      resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
      numberOfUsersAPThreshold: "20"
      callbackReference:
        notifyURL: "http://clientApp.example.com/location_notifications/123456"
      clientCorrelator: "0123"
      zoneId: "zone01"
      numberOfUsersZoneThreshold: "40"
  inline_response_200_notificationSubscriptionList:
    type: "object"
    properties:
      zonalTrafficSubscription:
        type: "array"
        items:
          $ref: "#/definitions/ZonalTrafficSubscription"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
  inline_response_200:
    type: "object"
    properties:
      notificationSubscriptionList:
        $ref: "#/definitions/inline_response_200_notificationSubscriptionList"
  inline_response_200_1_notificationSubscriptionList:
    type: "object"
    properties:
      userTrackingSubscription:
        type: "array"
        items:
          $ref: "#/definitions/UserTrackingSubscription"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
  inline_response_200_1:
    type: "object"
    properties:
      notificationSubscriptionList:
        $ref: "#/definitions/inline_response_200_1_notificationSubscriptionList"
  inline_response_200_2_notificationSubscriptionList:
    type: "object"
    properties:
      zonalTrafficSubscription:
        type: "array"
        items:
          $ref: "#/definitions/ZoneStatusSubscription"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
  inline_response_200_2:
    type: "object"
    properties:
      notificationSubscriptionList:
        $ref: "#/definitions/inline_response_200_2_notificationSubscriptionList"
...